Add files via upload
This was my second project in CS 401 Programming Languages at the University of Alabama at Birmingham.

The task is outlined in "a02-ruby.pdf" and as expected, students were expected to learn and utilize Ruby to complete this task. I worked with a partner on this assignment, so this is a joint effort.

Test cases are also included along with an assignment report.